+++ components/molecules/Button/Button.tsx
@@ -1,5 +1,5 @@
 import clsx from "clsx";
-import s from "./Button.module.sass";
+import s from "./style.module.sass";
 import Loader from "@/components/atoms/Loader/Loader";
 
 type Props = {
