+++ components/organisms/SignUpForm/style.module.sass
@@ -2,8 +2,9 @@
 
 .formContainer
   padding: 32px
-  border: 1px solid $gray
   +br8()
+  +shadow()
+  background-color: $white
 
   +viewS()
     padding: 16px
